[jaws.bar] 예약 대기 관리 step 1, 2 (#3)
* feat: auth 패키지 분리 및 의존성 최적화

* feat: 예약 대기 기능 구현

* test: 예약 대기 API 테스트 작성

* feat: GlobalExceptionHanlder 도입

- ErrorCode 및 ErrorResponse 도입
- Auth 패키지 내 AuthorizationException 도입
- GlobalExceptionHanlder 내 Auth 패키지 Exception 처리 추가

* refactor: Member Role Enum 적용

* feat: 요청값 Validation 구현

* refactor: Auth 패키지 Bean Configuration 클래스 분리

* style: 코드 리포맷팅

* feat: 예약 대기 중인 스케줄에 예약 대기 요청 시 예외 처리

- 해당 기능 Test 작성

* feat: 예약 취소 시 다음 순번 대기가 자동으로 예약되도록 기능 수정

- 수정된 기능에 대한 테스트 작성

* refactor: 피드백 반영

- auth 패키지 하위 세분화
- auth 패키지 내 컴포넌트 스캔용 어노테이션 제거
- null 검증 임시 획일화
- Transaction 적용
- 예약 대기 생성 로직 오류 수정

* refactor: critical section에 배타락 도입

- ReservationWaitingDao -> findByScheduleId
  - ReservationWaiting 생성 시 대기번호 산정 로직에 필요
  - Schedule 삭제 시 보유 예약 확인 로직에 필요
- ScheduleDao -> findByThemeId
  - Theme 삭제 시 보유 스케줄 확인 로직에 필요

* refactor: 트랜잭션 어노테이션 클래스에 적용

* refactor: 대기번호 산정 동기화 로직 변경

* feat: Lock 적용

- 예약 대기 번호 산정
- 스케줄 삭제 시 잔여 예약 검사
- 테마 삭제 시 잔여 스케줄 검사

* refactor: Lock 클래스 도입

+++ src/main/java/nextstep/schedule/ScheduleService.java
@@ -1,19 +1,36 @@
 package nextstep.schedule;
 
+import java.util.List;
+import java.util.Objects;
+import java.util.concurrent.atomic.AtomicInteger;
+import nextstep.common.Lock;
+import nextstep.error.ErrorCode;
+import nextstep.error.exception.RoomReservationException;
+import nextstep.reservation.dao.ReservationDao;
+import nextstep.reservation.dao.ReservationWaitingDao;
+import nextstep.reservation.domain.Reservation;
+import nextstep.reservation.domain.ReservationWaiting;
+import nextstep.reservation.service.ReservationService;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
+@Transactional
 public class ScheduleService {
     private ScheduleDao scheduleDao;
     private ThemeDao themeDao;
+    private ReservationDao reservationDao;
+    private ReservationWaitingDao reservationWaitingDao;
+    public static final Lock scheduleListLock = new Lock();
 
-    public ScheduleService(ScheduleDao scheduleDao, ThemeDao themeDao) {
+    public ScheduleService(ScheduleDao scheduleDao, ThemeDao themeDao, ReservationDao reservationDao,
+                           ReservationWaitingDao reservationWaitingDao) {
         this.scheduleDao = scheduleDao;
         this.themeDao = themeDao;
+        this.reservationDao = reservationDao;
+        this.reservationWaitingDao = reservationWaitingDao;
     }
 
     public Long create(ScheduleRequest scheduleRequest) {
@@ -21,11 +38,24 @@ public class ScheduleService {
         return scheduleDao.save(scheduleRequest.toEntity(theme));
     }
 
+    @Transactional(readOnly = true)
     public List<Schedule> findByThemeIdAndDate(Long themeId, String date) {
         return scheduleDao.findByThemeIdAndDate(themeId, date);
     }
 
     public void deleteById(Long id) {
+        Schedule schedule = scheduleDao.findById(id);
+        if (Objects.isNull(schedule)) {
+            throw new RoomReservationException(ErrorCode.SCHEDULE_NOT_FOUND);
+        }
+        ReservationService.reservationListLock.lock();
+        List<Reservation> reservationList = reservationDao.findByScheduleId(id);
+        List<ReservationWaiting> reservationWaitingList = reservationWaitingDao.findByScheduleId(id);
+        if (reservationList.size() > 0 || reservationWaitingList.size() > 0) {
+            ReservationService.reservationListLock.unlock();
+            throw new RoomReservationException(ErrorCode.SCHEDULE_CANT_BE_DELETED);
+        }
         scheduleDao.deleteById(id);
+        ReservationService.reservationListLock.unlock();
     }
 }
